Moved functions to path utils
