+++ utils/ui_utils.py
@@ -122,25 +122,3 @@ def fill_row(table: QTableWidget, row: int) -> None:
   add_checkbox(table, row, 13, checked=False)
   add_checkbox(table, row, 14, checked=False)
   add_checkbox(table, row, 15, checked=False)
-
-
-def discover_blender_path() -> None:
-  """Discover the path to Blender."""
-  possibilities = ["C:/Program Files (x86)/Steam/steamapps/common/Blender/blender.exe",
-                   "C:/Program Files/Blender Foundation/Blender/blender.exe",
-                   "/Applications/blender/blender.app/Contents/MacOS/blender",
-                   "/usr/bin/blender",
-                   "/usr/local/bin/blender",
-                   "~/blender/"]
-
-  for blender_path in possibilities:
-    if os.path.exists(blender_path):
-      return blender_path
-  return ""
-
-
-def get_blend_path(blend_path: str, blender_files_path: str) -> str:
-  """Get the absolute path of the blend file."""
-  if os.path.basename(blend_path) == blend_path:
-    return os.path.join(blender_files_path, blend_path)
-  return blend_path
